refactor: add hooks folder
