+++ src/app/components/SearchBox/SearchBox.tsx
@@ -3,7 +3,7 @@
 import React, { useCallback, useEffect, useState } from 'react';
 import { useRouter } from 'next/navigation';
 import Image from 'next/image';
-import useDebounce from '@/utils/useDebounce';
+import useDebounce from '@/hooks/useDebounce';
 import styles from './SearchBox.module.scss';
 
 function SearchBox() {
